Add -Wextra (and -Wno-unused-parameter)
There's a decent amount of functions that have unused parameters that I
consider those to be more or less safe (as they don't really matter).
The more worrisome parts are the many signed / unsigned comparisons.  I
have eliminated all the unsafe comparisons with this patch, and while
many of them probably weren't going to cause any trouble they all could
have been real bugs.

+++ src/pbashc/main.c
@@ -36,7 +36,7 @@ int main(int argc, char **argv)
     FILE *infile;
     FILE *outfile;
     char buffer[1024];
-    int read;
+    size_t read;
     char *chmod;
     int chmod_size;
